+++ TicketHub/__init__.py
@@ -5,4 +5,6 @@ async def setup(bot):
     cog = TicketHub(bot)
     if bot.get_command("ticket") is not None:
         cog.use_conflict_safe_prefix_root()
+    if bot.get_command("ticketset") is not None:
+        cog.use_conflict_safe_set_root()
     await bot.add_cog(cog)